Stellarium image generator/requirements.txt
anyio==4.9.0
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
djangorestframework==3.16.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
python-decouple==3.8
python-telegram-bot==20.7
requests==2.32.4
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
